requirements.txt
absl-py==1.4.0
alabaster==0.7.13
altair==5.1.1
altgraph==0.17.3
amqp==5.1.1
argh==0.28.1
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
bleach==6.0.0
blinker==1.6.2
bumpversion==0.5.3
cachetools==5.3.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.4
cmap==0.0.3
colorama==0.4.6
comet-ml==3.33.5
comm==0.1.3
comtypes==1.2.0
configobj==5.0.8
contourpy==1.1.0
coremltools==6.3.0
coverage==4.1
cryptography==41.0.3
customtkinter==5.2.0
cx-Freeze==6.15.5
cx-Logging==3.1.0
cycler==0.10.0
darkdetect==0.8.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
distlib==0.3.6
dlib==19.24.2
docu==0.24.0
docutils==0.20.1
dulwich==0.21.5
et-xmlfile==1.1.0
everett==3.1.0
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Faker==18.9.0
filelock==3.12.2
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
gast==0.4.0
gdown==4.7.1
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
gspread==5.11.0
gunicorn==20.1.0
h5py==3.9.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==2.10
ifaddr==0.2.0
imagesize==1.4.1
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kaggle==1.5.16
keras==2.13.1
keras-layer-normalization==0.16.0
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
kombu==5.2.4
kornia==0.7.0
lazy_loader==0.3
libclang==16.0.0
librosa==0.10.1
lief==0.13.2
llvmlite==0.41.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.3
mpmath==1.3.0
msgpack==1.0.7
mtcnn==0.1.1
mypy==1.2.0
mypy-extensions==1.0.0
neptune-tensorflow-keras==2.1.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.58.0
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
onnx==1.14.1
onnxsim==0.4.33
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.76
openpyxl==3.1.2
openvino==2023.0.1
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pefile==2023.2.7
photoshop-python-api==0.21.5
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.8.1
pluggy==1.2.0
pooch==1.7.0
prompt-toolkit==3.0.38
protobuf==4.24.3
psutil==5.9.5
psycopg==3.1.9
psycopg2==2.9.6
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyparsing==2.4.7
pypiwin32==223
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
PySocks==1.7.1
pytest==7.3.1
pytest-mock==3.10.0
pytest-qt==4.2.0
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pyvirtualcam==0.10.2
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0
pyzmq==25.1.0
QtAwesome==1.2.3
QtPy==2.3.1
redis==4.5.4
referencing==0.29.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retina-face==0.0.13
rich==13.4.2
roboflow==1.1.0
rpds-py==0.8.10
rsa==4.9
ruff==0.0.269
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
semantic-version==2.10.0
sentry-sdk==1.27.1
setproctitle==1.3.2
shiboken6==6.5.0
simplejson==3.19.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
snowballstemmer==2.2.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.6
Sphinx==1.4.8
SQLAlchemy==2.0.12
sqlparse==0.4.4
stack-data==0.6.2
supervision==0.11.1
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.5.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.12.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflowjs==3.18.0
termcolor==2.3.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tika==2.6.0
timm==0.9.2
toolz==0.12.0
torch==2.0.1
torch-tb-profiler==0.4.1
torchvision==0.15.2
tornado==6.3.2
tox==2.3.1
tqdm==4.65.0
traitlets==5.9.0
transparent-background==1.2.4
ttkthemes==3.2.2
ttkwidgets==0.13.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.154
urllib3==1.26.16
utils==1.0.1
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.3.3
wget==3.2
wrapt==1.15.0
wurlitzer==3.0.3
x2paddle==1.4.1
xknx==2.9.0